+++ frontend/src/components/CartItems/CartItems.tsx
