+++ frontend/src/components/MainLayout/components/Cart.tsx
