+++ frontend/src/constants/apiPaths.ts
@@ -2,7 +2,7 @@ const API_PATHS = {
   product: "https://olek-shop-app.s3-website-eu-west-1.amazonaws.com/dev",
   order: "https://olek-shop-app.s3-website-eu-west-1.amazonaws.com/dev",
   import: "https://olek-shop-app.s3-website-eu-west-1.amazonaws.com/dev",
-  bff: "https://olek-shop-app.s3-website-eu-west-1.amazonaws.com/dev",
+  bff: "https://7dnbglzgak.execute-api.eu-west-1.amazonaws.com/dev",
   cart: "https://olek-shop-app.s3-website-eu-west-1.amazonaws.com/dev",
 };
