+++ frontend/src/models/CartItem.ts
