+++ frontend/src/models/Product.ts
