--- src/index.tsx
@@ -1,35 +1,0 @@
-import React from "react";
-import { createRoot } from "react-dom/client";
-import App from "~/components/App/App";
-import CssBaseline from "@mui/material/CssBaseline";
-import { ThemeProvider } from "@mui/material/styles";
-import { BrowserRouter } from "react-router-dom";
-import { QueryClient, QueryClientProvider } from "react-query";
-import { ReactQueryDevtools } from "react-query/devtools";
-import { theme } from "~/theme";
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: { refetchOnWindowFocus: false, retry: false, staleTime: Infinity }
-  }
-});
-
-import("./mocks/browser").then(({ worker }) => {
-  worker.start({ onUnhandledRequest: "bypass" });
-  const container = document.getElementById("app");
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const root = createRoot(container!);
-  root.render(
-    <React.StrictMode>
-      <BrowserRouter>
-        <QueryClientProvider client={queryClient}>
-          <ThemeProvider theme={theme}>
-            <CssBaseline />
-            <App />
-          </ThemeProvider>
-          <ReactQueryDevtools initialIsOpen={false} />
-        </QueryClientProvider>
-      </BrowserRouter>
-    </React.StrictMode>
-  );
-});
